+++ src/bindings/bnd_polyline.h
@@ -33,9 +33,11 @@ public:
   void SetAllX(double xValue);
   void SetAllY(double yValue);
   void SetAllZ(double zValue);
-  void Append(int count, const std::vector<ON_3dPoint>& points) { m_polyline.Append(count, points); }
+  
   #if defined(ON_PYTHON_COMPILE)
-  void Append2(int count, const std::vector<ON_3dPoint>& points)
+  void Append2(int count, pybind11::object points);
+  #else
+  void Append(int count, const std::vector<ON_3dPoint>& points) { m_polyline.Append(count, points); }
   #endif
   //public static int ClosestIndexInList(IList<Point3d> list, Point3d testPoint)
   //public static Point3d ClosestPointInList(IList<Point3d> list, Point3d testPoint)
@@ -74,5 +76,5 @@ public:
   static BND_Polyline* CreateInscribedPolygon(class BND_Circle& circle, int sideCount);
   static BND_Polyline* CreateCircumscribedPolygon(class BND_Circle& circle, int sideCount);
   static BND_Polyline* CreateStarPolygon(class BND_Circle& circle, double radius, int cornerCount);
-  static BND_Polyline* CreateFromPoints(const std::vector<ON_3dPoint>& points)
+  static BND_Polyline* CreateFromPoints(const std::vector<ON_3dPoint>& points);
 };